style: move FirestoreDatabaseService to service.go
